# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title:  DevOps Ecosystem Platform — Combined API
  version: "2.0.0"
  description: |
    API contract for 4 modules added in BRD-001 v1.1 (FR-019 – FR-032).
    Auth: Bearer JWT (ADR-003). All write operations produce audit log (NFR-007).
    Banking-grade: all endpoints except /health require Authorization header.
  contact:
    name: Backend Dev (Agent 06)
servers:
  - url: http://localhost:8001
    description: PPG System — Annual Plans, Project Objects, Reports
  - url: http://localhost:8002
    description: BA Workflow — BA Documents v2
  - url: http://localhost:8003
    description: Test Platform — Test Documents
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/documents:
    get:
      summary: List BA documents (FR-027, FR-028, FR-029)
      operationId: listBADocuments
      tags: [ba-documents]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: doc_type
          in: query
          schema:
            type: string
            enum: [BRD, BRS, FSD, API_SPEC, ERD, DATA_DICT, WIREFRAME, PROCESS_FLOW]
        - name: status
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by linked object (FR-027)
        - name: milestone_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of BA documents

    post:
      summary: Create BA document (FR-027, FR-029)
      operationId: createBADocument
      tags: [ba-documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, doc_type, title]
              properties:
                project_id:
                  type: string
                  format: uuid
                doc_type:
                  type: string
                  enum: [BRD, BRS, FSD, API_SPEC, ERD, DATA_DICT, WIREFRAME, PROCESS_FLOW]
                title:
                  type: string
                  maxLength: 300
                content:
                  type: string
                  default: ""
                milestone_id:
                  type: string
                  format: uuid
                  nullable: true
                object_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                metadata:
                  type: object
      responses:
        "201":
          description: Created document
        "422":
          description: OBJECT_INCOMPLETE or validation error (BR-011)
  /api/v1/documents/{doc_id}/status:
    post:
      summary: Change document status (FR-009, BR-001)
      operationId: changeBADocStatus
      tags: [ba-documents]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [submit_review, approve, reject, archive]
                notes:
                  type: string
      responses:
        "200":
          description: Updated status. If BRS approved, triggers test case auto-gen.
        "409":
          description: STATE_MACHINE_VIOLATION
...
  /api/v1/test-documents:
    get:
      summary: List test documents (FR-030, FR-031, FR-032)
      operationId: listTestDocuments
      tags: [test-documents]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: doc_type
          in: query
          schema:
            type: string
            enum: [TEST_PLAN, BUG_REPORT, UAT_SIGNOFF]
        - name: status
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by linked object (FR-030)
        - name: milestone_id
          in: query
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
          description: BUG_REPORT only
      responses:
        "200":
          description: List of test documents

    post:
      summary: Create test document (FR-032)
      operationId: createTestDocument
      tags: [test-documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, doc_type, title]
              properties:
                project_id:
                  type: string
                  format: uuid
                doc_type:
                  type: string
                  enum: [TEST_PLAN, BUG_REPORT, UAT_SIGNOFF]
                title:
                  type: string
                content:
                  type: string
                object_id:
                  type: string
                  format: uuid
                  nullable: true
                milestone_id:
                  type: string
                  format: uuid
                  nullable: true
                metadata:
                  type: object
                  description: |
                    BUG_REPORT: {severity: critical|high|medium|low}
                    UAT_SIGNOFF: {scope: string}
      responses:
        "201":
          description: Created test document
        "422":
          description: Invalid doc_type or invalid severity
  /api/v1/test-documents/{doc_id}/status:
    post:
      summary: Change test document status per type state machine (FR-032)
      operationId: changeTestDocStatus
      tags: [test-documents]
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  description: Depends on doc_type state machine
                notes:
                  type: string
                approver:
                  type: string
                  description: Required for UAT_SIGNOFF sign action
                sign_date:
                  type: string
                  format: date
                  description: Required for UAT_SIGNOFF sign action
      responses:
        "200":
          description: Updated status
        "409":
          description: STATE_MACHINE_VIOLATION
...
  /api/v1/test-documents/objects/{object_id}/test-coverage:
    get:
      summary: Test coverage per object with milestone alerts (FR-031)
      operationId: getObjectTestCoverage
      tags: [test-documents]
      parameters:
        - name: object_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: |
            Coverage summary: total_test_cases, executed, passed, failed,
            coverage_pct, milestone_coverage (with alert when below threshold)
...